+++ checkstyle/suppressions.xml
@@ -75,7 +75,6 @@
               files="com/hazelcast/cache/impl/AbstractCacheRecordStore"/>
     <suppress checks="MethodCount" files="com/hazelcast/cache/impl/AbstractCacheService"/>
     <suppress checks="MethodCount" files="com/hazelcast/cache/impl/nearcache/impl/store/AbstractNearCacheRecordStore"/>
-    <suppress checks="NPathComplexity" files="com/hazelcast/cache/impl/eviction/impl/evaluator/AbstractEvictionPolicyEvaluator"/>
     <suppress checks="MethodCount" files="com/hazelcast/cache/impl/AbstractHazelcastCacheManager"/>
     <suppress checks="ClassFanOutComplexityCheck" files="com/hazelcast/cache/impl/AbstractCacheService"/>
