+++ hazelcast/src/main/java/com/hazelcast/internal/eviction/EvictionListener.java
@@ -14,24 +14,27 @@
  * limitations under the License.
  */
 
-package com.hazelcast.cache.impl.eviction;
+package com.hazelcast.internal.eviction;
 
 /**
  * Interface for entries, records or whatever that can be evicted via its accessor (key or id).
  *
  * @param <A> Type of the accessor
- * @param <E> Type of the {@link com.hazelcast.cache.impl.eviction.Evictable} value
+ * @param <E> Type of the {@link Evictable} value
  */
 public interface EvictionListener<A, E extends Evictable> {
 
+    /**
+     * Empty {@link} EvictionListener
+     */
     EvictionListener NO_LISTENER = null;
 
     /**
-     * Called when an {@link com.hazelcast.cache.impl.eviction.Evictable} entry is evicted.
+     * Called when an {@link Evictable} entry is evicted.
      *
-     * @param evictedEntryAccessor  Accessor of the {@link com.hazelcast.cache.impl.eviction.Evictable} entry
+     * @param evictedEntryAccessor  Accessor of the {@link Evictable} entry
      *                              that is evicted.
-     * @param evictedEntry          {@link com.hazelcast.cache.impl.eviction.Evictable} entry that is evicted.
+     * @param evictedEntry          {@link Evictable} entry that is evicted.
      */
     void onEvict(A evictedEntryAccessor, E evictedEntry);
